Refactored assign class
